requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.3
pandas==2.2.2
python-dotenv==1.0.1
pyinstaller==6.8.0
webdriver-manager==4.0.2
requests==2.31.0
lxml==5.2.2